+++ errors/__init__.py
@@ -2,6 +2,7 @@ from .base_bot_error	import BaseBotError
 
 from .pi_prompt_busy_error			import PIPromptBusyError
 from .session_manager_closing_error	import SessionManagerClosingError
+from .miss_factory_error			import MissFactoryError
 
 
 __all__ = [
@@ -11,5 +12,6 @@ __all__ = [
 	
 	"PIPromptBusyError",
 	"SessionManagerClosingError",
+	"MissFactoryError",
 
 ]