+++ errors/miss_factory_error.py
@@ -1,0 +1,7 @@
+from .base_bot_error	import BaseBotError
+
+
+class MissFactoryError(BaseBotError):
+	
+	"""我是doc(不知道写啥"""
+	...