+++ protocols/__init__.py
@@ -4,12 +4,17 @@ from .store	import StoreProtocol
 
 # 运行时检查
 from .parser	import Parser
+from .closable	import Closable
 
 
 __all__ = [
 	
+	# abc
 	"ChainProtocol",
 	"StoreProtocol",
+	
+	# runtime
 	"Parser",
+	"Closable",
 	
 ]