+++ protocols/closable.py
@@ -1,0 +1,7 @@
+from typing	import Protocol, runtime_checkable
+
+
+@runtime_checkable
+class Closable(Protocol):
+	
+	async def close(self) -> None: ...