+++ protocols/parser.py
@@ -1,6 +1,7 @@
-from typing	import Protocol, Any
+from typing	import Protocol, Any, runtime_checkable
 
 
+@runtime_checkable
 class Parser(Protocol):
 		
 	async def is_accept(self, data: Any) -> bool: ...